README.md
Knowledge is a Learning Management System for blended learning, distance education, flipped classroom and other eLearning projects. eKnowledge is an online learning platform, which is mobile based as we understand that not every highschool learner has a laptop or a computer. However, we know that a majority if not all have smartphones.  Moreover, all of the data can be updated via our backend system, providing dynamic and real-time data.
It allows one to provide dynamic content to students in a central location, communicate with students quickly, and provide grades in an electronic format to students.  It has all the tools that the student needs, from scientific calculator, maths problem solver, system of equations, 2d geometry, financial maths from equated monthly installments, compound interest, loan affordability, simple interest and many more. A student can create a custom time-table, organise upcoming events such as upcoming homework/test, create a schedule when a homework or assignment is due, they can write notes, write assignments and submit assignments through the App. When a student fills up the attendance register or submit an assignment or engages with the App, everything is stored to the cloud database with a time stamp. The App is using machine learning such as OCR optical character recognition and NLP natural language processing. When a student submit an assignment or engages with the app, since everything is connected to our backend system, we can manipulate or modify data from our end, and eveything happens in real time. Furthermore, the institution the student has enrolled at provides the data, data of which they want a specific grade to see or the entire school or the specific student,  data of which they can provide can be documents such as textbooks, study-guides, upload course content such as lessons, video tutorials and many more. 
It brings  an enhcaned user experience with high-quality, interactive video chat. Our platform gives you the building blocks you need to assemble interactive classrooms anywhere, on any device, for any class size.
eKnowledge provides everything you would expect in a state-of-the-art virtual classroom—and many of the unexpected: AI assistance, fun features and a global network that reaches even small villages with challenging access to education. 
eKnowledge is designed to provide educators, administrators, and learners with an open, robust, secure platform to create and deliver personalized learning environment. Nevertheless, it delivers a powerful set of learner-centric tools and collaborative learning environments that empower both teaching and learning.